+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:39823/98cca2b6-c9bf-44fc-bbf4-45a0a71a5da0?"}
+{"mongoUri":"mongodb://127.0.0.1:43385/91782eea-c1df-458a-8705-1014587a0b4f?"}
