+++ src/main/factories/login/login-validation-factory.ts
@@ -1,6 +1,6 @@
 import { RequiredFielValidation, EmailValidation, ValidationComposite } from '../../../presentation/helpers/validators'
 import { Validation } from '../../../presentation/protocols/validation'
-import { EmailValidatorAdapter } from '../../../utils/email-validator-adapter'
+import { EmailValidatorAdapter } from '../../adapters/validators/email-validator-adapter'
 
 export const makeLoginValidation = (): Validation => {
   const validations: Validation[] = []
