+++ src/main/factories/signup/signup-validation.ts
@@ -5,7 +5,7 @@ import {
   EmailValidation
 } from '../../../presentation/helpers/validators'
 import { Validation } from '../../../presentation/protocols/validation'
-import { EmailValidatorAdapter } from '../../../utils/email-validator-adapter'
+import { EmailValidatorAdapter } from '../../adapters/validators/email-validator-adapter'
 
 export const makeSignUpValidation = (): Validation => {
   const validations: Validation[] = []
